Modularize Gradio app into separate views and compose in app.py
